так.. коннект почти налажен.. надо разобраться с архивацией (пароль), вероятно, другой нугет попробую, пока не получается скачать архив и чтобы он был валидный. после того, как коннект закончу, буду его тестировать на АНДРОИДЕ

+++ Dto/FileObj.cs
@@ -14,7 +14,10 @@
 
         public void Dispose()
         {
-            File.Delete(_name);
+            if (File.Exists(_name))
+            {
+                File.Delete(_name);
+            }
         }
     }
 }
